+++ packages/components/src/Button/Button.tsx
@@ -1,23 +1,20 @@
 import { forwardRef } from 'react';
-import { dj } from '@danji/styled';
+import { dj, useThemeStyles } from '@danji/styled';
 import { ButtonOptions } from './button.types';
-import Components from '../theme';
 
 const Button = forwardRef<HTMLButtonElement, ButtonOptions>((props, ref) => {
   const { children, type = 'primary', variant = 'solid', size = 'md' } = props;
 
-  const baseStyleProps = { type, variant, size };
-
-  const sizeStyles = Components.Button.sizes[baseStyleProps.size];
-  const colorStyles = Components.Button.variants[baseStyleProps.variant](
-    baseStyleProps.type,
-  );
-
-  const styles = {
+  const baseStyles = {
     borderRadius: '0.325rem',
     cursor: 'pointer',
-    ...sizeStyles,
-    ...colorStyles,
+  };
+
+  const buttonThemeStyles = useThemeStyles('Button', { type, variant, size });
+
+  const styles = {
+    ...baseStyles,
+    ...buttonThemeStyles,
   };
 
   return (
