+++ packages/components/src/index.ts
@@ -1,1 +1,2 @@
 export * from './Button';
+export { default as Theme } from './theme';
