+++ packages/styled/src/theme/hooks.ts
@@ -1,0 +1,29 @@
+import { ThemeContext } from '@emotion/react';
+import { useContext } from 'react';
+import { ThemeWithCssVars } from '../provider.types';
+import { getValueByPath } from './base';
+
+const useTheme = <T extends Record<string, any>>() => {
+  const ctx = useContext(ThemeContext);
+
+  if (ctx === null || ctx === undefined) throw new Error('useTheme error');
+
+  return ctx as ThemeWithCssVars<T>;
+};
+
+export const useThemeStyles = (themeKey: string, props: any) => {
+  const theme = useTheme();
+  const { variant, size, type } = props;
+  const styles = {};
+
+  const themeStyleConfig = getValueByPath(theme, `components.${themeKey}`);
+
+  if (themeStyleConfig) {
+    Object.assign(styles, {
+      ...themeStyleConfig.sizes[size],
+      ...themeStyleConfig.variants[variant](type),
+    });
+  }
+
+  return styles;
+};
