+++ packages/styled/src/theme/index.ts
@@ -1,0 +1,3 @@
+export * from './base';
+export * from './cssVar';
+export * from './hooks';
